// frontend/src/api-client/model/quiz-analytics-session-entity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Quiziset API
 * The RESTful APIs Docs for the Quiziset app
 * OpenAPI spec version: 1.0
 */
import type { PublicUserEntity } from './public-user-entity'

export interface QuizAnalyticsSessionEntity {
  id: number
  /** the user who took the quiz */
  user: PublicUserEntity
  /** total number of questions in the session */
  questionsCount: number
  /** number of correctly answered questions */
  successfulAnswersCount: number
  startTime: string
  /**
   * null while the session is still ongoing
   * @nullable
   */
  finishTime: string | null
}